+++ municipal-services/tl-services/src/main/java/org/egov/tl/repository/TLRepository.java
@@ -10,6 +10,7 @@ import org.egov.tl.web.models.*;
 import org.egov.tl.workflow.WorkflowService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.SingleColumnRowMapper;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.CollectionUtils;
 
@@ -127,5 +128,26 @@ public class TLRepository {
         });
     }
 
+    public List<TradeLicense> getPlainLicenseSearch(TradeLicenseSearchCriteria criteria) {
+        List<Object> preparedStmtList = new ArrayList<>();
+        String query = queryBuilder.getTLPlainSearchQuery(criteria, preparedStmtList);
+        log.info("Query: " + query);
+        List<TradeLicense> licenses =  jdbcTemplate.query(query, preparedStmtList.toArray(), rowMapper);
+        sortChildObjectsById(licenses);
+        return licenses;
+    }
+
+    public List<String> fetchTradeLicenseIds(TradeLicenseSearchCriteria criteria){
+
+        List<Object> preparedStmtList = new ArrayList<>();
+        preparedStmtList.add(criteria.getOffset());
+        preparedStmtList.add(criteria.getLimit());
+
+        return jdbcTemplate.query("SELECT id from eg_tl_tradelicense ORDER BY createdtime offset " +
+                        " ? " +
+                        "limit ? ",
+                preparedStmtList.toArray(),
+                new SingleColumnRowMapper<>(String.class));
+    }
 
 }