+++ municipal-services/tl-services/src/main/java/org/egov/tl/service/notification/PaymentNotificationService.java
@@ -77,6 +77,7 @@ public class PaymentNotificationService {
 
     final String receiptNumberKey = "receiptNumber";
 
+    final String payerNameKey = "payerName";
 
     /**
      * Generates sms from the input record and Sends smsRequest to SMSService
@@ -216,7 +217,7 @@ public class PaymentNotificationService {
         else
             message = util.getPayerPaymentMsg(license,valMap,localizationMessages);
 
-        String customizedMsg = message.replace("<1>",valMap.get(paidByKey));
+        String customizedMsg = message.replace("<1>",valMap.get(payerNameKey));
         SMSRequest smsRequest = new SMSRequest(valMap.get(payerMobileNumberKey),customizedMsg);
         return smsRequest;
     }
@@ -239,10 +240,11 @@ public class PaymentNotificationService {
             valMap.put(businessServiceKey,businessServiceList.isEmpty()?null:businessServiceList.get(0));
             valMap.put(consumerCodeKey,consumerCodeList.isEmpty()?null:consumerCodeList.get(0));
             valMap.put(tenantIdKey,context.read("$.Payment.tenantId"));
-            valMap.put(payerMobileNumberKey,mobileNumberList.isEmpty()?null:mobileNumberList.get(0));
+            valMap.put(payerMobileNumberKey,context.read("$.Payment.mobileNumber"));
             valMap.put(paidByKey,context.read("$.Payment.paidBy"));
             valMap.put(amountPaidKey,amountPaidList.isEmpty()?null:String.valueOf(amountPaidList.get(0)));
             valMap.put(receiptNumberKey,receiptNumberList.isEmpty()?null:receiptNumberList.get(0));
+            valMap.put(payerNameKey,context.read("$.Payment.payerName"));
         }
         catch (Exception e){
             e.printStackTrace();