+++ municipal-services/tl-services/src/main/java/org/egov/tl/web/controllers/TradeLicenseController.java
@@ -88,6 +88,17 @@ import javax.servlet.http.HttpServletRequest;
         return new ResponseEntity(HttpStatus.ACCEPTED);
     }
 
+    @RequestMapping(value="/_plainsearch", method = RequestMethod.POST)
+    public ResponseEntity<TradeLicenseResponse> plainsearch(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
+                                                            @Valid @ModelAttribute TradeLicenseSearchCriteria criteria){
+
+        List<TradeLicense> licenses = tradeLicenseService.plainSearch(criteria,requestInfoWrapper.getRequestInfo());
+
+        TradeLicenseResponse response = TradeLicenseResponse.builder().licenses(licenses).responseInfo(
+                responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(), true))
+                .build();
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }